Move preview buttons to separate file
